Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Polyfills should be provided by the user. Please move this to the other polyfill in wasm_exec.html.
> 
> Who is the author of this polyfill? Are we allowed to use it? Do we need to add attribution?

The author is the GitHub user Yaffle, and it was taken from their gists here:
https://gist.github.com/Yaffle/5458286

Since it doesn't seem to have a license attached, I'll assume we cannot use it.

I tested the following polyfill and that seems to work fine as well, it's also licensed under "Apache License 2.0" which seems like it should be OK? (Though I'm not sure when it comes to licensing)
https://github.com/samthor/fast-text-encoding/

I also had a look for other implementations on GitHub and looked at their licensing, but it seemed to either use "Apache License 2.0", "Creative Commons Attribution 4.0 International License" or be unclear.
ie.
- https://github.com/inexorabletash/text-encoding
- https://github.com/coolaj86/TextEncoderLite

If this is a problem, is it OK if I rewrite the implementation so that it doesn't resemble these?
Alternatively, I could just change this PR to simply be a code comment that tells the user that if they want Edge support, they will need a polyfill for TextEncoder/TextDecoder.

Thoughts?

Patch-set: 1
Reviewer: Gerrit User 28917 <28917@62eb7196-b449-3ce5-99f1-c037f21e1705>