net: extend documentation for net.FileConn()
On all supported platforms, net.FileConn does a syscall to look up what
type of socket it's been passed then returns the appropriate
implementation of net.Conn.

Make the documentation for this function clear about this behavior, as
it was previously not obvious from the description that there wasn't a
special private implementation of net.Conn that it would return.

Also extend net.TestFileConn to enforce that the extended doc-string is
not lying.

Add a shorter statement to the net.Dial documentation string to make it
clear that it isn't returning a private struct either.

fixes: #27391

Change-Id: I40b780e19cc6b76b7c56dd32e52a1461beebe33a