net: extend documentation for net.FileConn()
On all supported platforms, net.FileConn does a syscall to look up what
type of socket it's been passed then returns the appropriate
implementation of net.Conn.

Make the documentation for this function clear about this behavior, as
it was previously not obvious from the description that there wasn't a
special private implementation of net.Conn that it would return.

Also extend net.TestFileConn to enforce that the extended doc-string is
not lying.

Reformat the list of valid networks in the net.Dial documentation string
as a table to make it what's returned for each network.

fixes: #27391

Change-Id: I40b780e19cc6b76b7c56dd32e52a1461beebe33a