Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (4 comments)

I've made the changes requested.
I've also moved the polyfill to execute before "wasm_exec.js" because `TextEncoder` and `TextDecoder` are referenced immediately when the script is included, causing a crash in Edge.

Perhaps the encoder/decoder const variables in wasm_exec.js should be instantiated in the Go class constructor or similar? (So the polyfill can be included after the script but before initialization of `new Go()`)

Patch-set: 3