Update patch set 5
Patch Set 5:

I'm not a big fan of these examples: they show the API of sync.Map, but don't illustrate how to use it effectively. But the API is the simple part — the harder part is to decide whether sync.Map is appropriate.

The documentation for sync.Map says “[m]ost code should use a plain Go map instead […] for better type safety and to make it easier to maintain other invariants along with the map content.” So the examples should illustrate exceptions to that rule: cases that require concurrency, and are either write-once/read-many or involve concurrent updates to disjoint (but stable) keys.

Patch-set: 5
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>