Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > The test is wrong, it should test that AND is not emitted; ORR was emitted even before this CL.
> 
> Thank you for your review. Because I am not familiar with codegen test framework. Adding "// arm64:"ORR\t[$]-9223372036854775808", -"AND\t[$]9223372036854775807"   (no and operand)"  comments is right test? Thank you.

I don't think the exact numbers are helpful here (if they were wrong, I assume that a more specific test of the functionality of math.Copysign would fail elsewhere). What you need to test in the codegen test framework if the rule you just added triggers. The goal of the rule is to remove the superfluous AND call and just leave ORR. Doing:

    // arm64:"ORR",-"AND"

seems enough to me.

Patch-set: 4