Update patch set 2
Patch Set 2:

> > Do we need a specific flag for this?  The linker presumably has to understand that it is working with MSVC files; can it simply automatically do the right thing in that case?
> 
> The thought process here was that in the future when support for the MSVC version of clang is added to Go (on my to do list) that the external flag will make it easier to decide what should happen given clang can behave as gcc or msvc on windows.

Even then, can't we test the compiler, rather than requiring a flag?

Or perhaps the go tool is going to do the test?  I suppose it's not so bad if the user is never required to use the flag themselves.

Patch-set: 2
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>