doc: make golang-nuts discussion list more prominent
The discussion list was buried beneath the developer mailing list.
The website is most likely be used by people unfamiliar with
the Go project. This puts the discussion list at the top of the page
so it is easy to find. This might reduce the number of misdirected
questions that are sent to golang-dev.

Change-Id: I8dcb4af98e454ae3a0140f9758a5656909126983