time: make TZ parse more tolerant
According to POSIX, there are three formats for TZ variable. When
it refers to timezone file, it should starts with a colon. This commit
removes the colon if it exists, so that it keeps compatible with both
the spec and the old behavior.

Change-Id: I30cfeaea530d24e174de309952338cb1146694a5
GitHub-Last-Rev: d7779a544f5a943f60d0cd23c860dabcc8e7364b
GitHub-Pull-Request: golang/go#27570