os/user: retrieve username from /etc/passwd instead of $USER
Per https://github.com/golang/go/issues/27524 there are situations
where the username for the uid does not match the value in the $USER
environment variable and it seems sensible to choose the value in
/etc/passwd when they disagree.

This may make the Current() call slightly more expensive, since we
read /etc/passwd with cgo disabled instead of just checking the
environment. However, we cache the result of Current() calls, so we
only invoke this cost once in the lifetime of the process.

Fixes #27524.

Change-Id: I0dcd224cf7f61dd5292f3fcc363aa2e9656a2cb1