Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > > This doesn't look correct.  TZ=:/my/favorite/timezone should read exactly the file /my/favorite/timezone, but this will instead search for that file in various system directories.
> > > 
> > > Also this should have a test.
> > 
> > Yes. This PR is not a fully implementation of the spec. It just tries to make TZ accept a `:` prefix. Absolute path and raw offset configurations are not supported (just like before).
> 
> But I think it's not just that they are not supported, as far as I can tell ":/America/Los_Angeles" will actually do the wrong thing: it will return success when it should not.

Do you mean the it should fatal when error is found? In the original implement, Local falls back to UTC when error is detected. For example, when setting TZ to "/America/Los_angeles", Local will be UTC.

And I'm not sure if it's appropriate to fatal inside the initial function.

Patch-set: 3