Update patch set 4
Patch Set 4:

(3 comments)

Thanks. It seems to me that the generated test file is trying too hard to be "real" while at the same time it's not (and won't compile). The only thing that matters is that is parses, and the size of the generated variable initializer, which dominates processing time.

I'd reduce it to exactly that: a minimal package with huge variable initializers. Then the benchmark is very clear.

There's a second part, the bindata map, which seems responsible for roughly one half of the allocations. I'd make that a separate benchmark, again, reduced to minimum.

That way this file becomes shorter and easier to understand, and the benchmark becomes clearer.

I don't see much benefit in trying to make the generated code as "real" as possible - it only distorts the benchmark by adding extraneous stuff.

Patch-set: 4
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>