Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> Hi Roustem,
> 
> As Robert mentioned above this is a change to the standard library. Any change of this nature should be discussed first in an issue, https://golang.org/issue/new, and a decision to proceed or not reached before sending a change.
> 
> As it stands now this change will not be accepted. I encourage you to open an issue, https://golang.org/issue/new to discuss the experience your developers had that lead you to propose this change, but understand that the solution may not be to change the container types, but instead to change the surrounding code to not use this pattern.
> 
> Thanks
> 
> Dave

I agree with Robert that the pattern of using *list.List pointer is unusual and should not be encouraged. It seems that it would be simpler to get rid of the pointer on our side.

I am going to close the PR. Thank you for your time!

Roustem

Patch-set: 1