encoding/json: resolve struct, map, array, or slice values in interface types
This work extends the behaviour of `json.Unmarshal` to allow for setting into concrete types
embedded in `interface{}` struct fields. For example, the following JSON object:

    {"A": {"B": "C"}}

Being unmarshaled into the following structure:

	var t = struct{ A interface{} }{A: struct{ B string }{}}

Would have previously resulted in field `A` being replaced with a `map[string]interface{}`,
disregarding the underlying type given to the `interface{}` field, i.e.:

    struct { A interface {} }{A:map[string]interface {}{"B":"C"}}

Whereas, by resolving the underlying type of the `interface{}` field, we are able to set into and
return the correct type, i.e.:

    struct { A interface {} }{A:struct { B string }{B:"C"}}

These rules apply also to `map` types, for JSON objects, and arrays or slices, for JSON arrays. The
rules for setting into these fields are exactly the same as those that would apply to the bare types
themselves, i.e. attempting to unmarshal into an underlying type that is unable to store the values
presented in the JSON input will return a type error.

Noted that this is a breaking change for users that relied on having set a concrete type, but handle
a `map[string]interface{}` or `[]interface{}` on the other side.

Fixes #26946

Change-Id: I207c984df4f53ac7c232119fe0760738750f0a87
GitHub-Last-Rev: 0af936806369a7062d40f39c9d89dbc677eb9c86
GitHub-Pull-Request: golang/go#27701