Update patch set 1
Patch Set 1:

It seems to me like this should be discussed on the issue first - it's a non-trivial change in behavior that could lead to subtly breaking existing programs.

I realise it hasn't gotten much attention, but that's what the NeedsDecision label is for (as well as proposals, for more complex things).

Patch-set: 1