Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It seems to me like this should be discussed on the issue first - it's a non-trivial change in behavior that could lead to subtly breaking existing programs.
> 
> I realise it hasn't gotten much attention, but that's what the NeedsDecision label is for (as well as proposals, for more complex things).

Correct, I've noted as much on the commit message and the initial issue on Github. Even if the modified behaviour here is considered to be "correct", it's gonna be hard to catch in 'go vet', and potentially hard to debug.

I mainly did the work here as the code ended up being fairly simple, and as a way of pushing to some sort of consensus. If changing the existing behaviour is considered unacceptable, I could look into clarifying the documentation on json.Unmarshal at the very least.

Patch-set: 1
CC: Gerrit User 29237 <29237@62eb7196-b449-3ce5-99f1-c037f21e1705>