encoding/json: fix usage of sync.Pool
The current usage of sync.Pool is leaky because it stores an arbitrary
sized buffer into the pool. However, sync.Pool assumes that all items in the
pool are interchangeable from a memory cost perspective. Due to the unbounded
size of a buffer that may be added, it is possible for the pool to eventually
pin arbitrarily large amounts of memory in a live-lock situation.

As a simple fix, we just set a maximum size that we permit back into the pool.
A more complex fix could use tiers of pools, which are tailored to specific
size classes.

Fixes #27735
Updates #23199

Change-Id: If3f19de0fa46a5fd67df615670a56f6ae86fe37d