sync: clarify proper Pool usage for dynamically sized buffers
A Pool is not suitable for use with items that may be backed by arbitrarily
large buffers. In a live-lock situation, large buffers may be reused indefinitely
causing a form of memory leak.

This changes the following:
* Remove the suggestion of following fmt's example use of Pools,
which also suffers from a live-lock memory leak.
* Fix the example by adding a simple upper-bound threshold for items being
returned to the pool.

Fixes #23199

Change-Id: I34a36e91c61d5967425d2db6245b5de6bceddb78