cmd/go/internal/invariant: collect invariant violations
Sample output:

    ~/go/src$ go build -o failgo cmd/go && ./failgo
    internal error at /home/bcmills/go/src/cmd/go/main.go:87: Invariants should not fail intentionally.
    Please report this, ideally with steps to reproduce, at https://golang.org/issue.

    Go is a tool for managing Go source code.
    […]

In previous code reviews, I have suggested that we add panics to detect
invariant violations during testing, and Russ has sometimes been reluctant to
add them out of concern that such panics may needlessly fail commands that could
otherwise succeed.

This change adds an experimental package for collecting invariant violations.
Its behavior can be customized via a hook. That way, we can panic
during tests, but only log violations (and return an explicit boolean to the
caller for recovery) in the production code path.

The fact that the hook can be customized makes the invariant package
intentionally useless for control flow: unlike a panic, an invariant cannot be
recovered reliably.

The fact that the hook can return control to the caller helps to mitigate
https://golang.org/doc/faq#assertions: since the hook may merely log the
violation and return to the caller, they are still prompted to consider how the
program should proceed in case of error.

If this package proves useful and usable, I intend to turn it into a proposal
for standard library integration. I believe it is a candidate for the standard
library because there should be only one customization point per binary, the
testing package should make use of that customization point automatically for
test binaries, and the compiler should synthesize descriptions for simple
conditions rather than requiring an explicit string in user code.

Change-Id: Iffaa17aba2e3319ee9a9d19b6386e7e99c96be38