Update patch set 1
Patch Set 1:

I'd drop the Note prefix. All docs are notes. Also, the hyphen instead of a colon is weird. And we can probably omit the final EUID jargon acronym. It doesn't add anything for people who understand the implications from the caching part. And if they don't know what EUID etc mean, then the docs only serve to confuse if you leave it in.

Patch-set: 1
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>