Update patch set 2
Patch Set 2:

LGTM overall.

One question that occurs to me: what will happen for global data symbols that are imported as a result of a call to "hostArchive()" -- is this something you have tested?  This is tricky to test on Linux since the only applicable host archive is libgcc.

Patch-set: 2
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>