Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I already submitted https://go-review.googlesource.com/c/go/+/137856 yesterday.
> 
> Note that this CL probably won't build with a Go 1.4 bootstrap.

If that's true, let's not do this. I use Go 1.4 to bootstrap and this change is not a compelling reason to rebuild my installation. If bootstrapping is to require a later version of the installation to run in future, that change should be planned and discussed not an unintended consequence of a CL to tidy up some code with no semantic effect.

Patch-set: 2
CC: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>