cmd: fix symbols access for AIX
This commit changes how symbols are accessed on AIX operating system.

It allows ppc64 to have symbols on 64 bits addresses, by adding two
instructions before the "linux" symbols access (on 32 bits). A new
relocation type is added for these two instructions and will relocate
bits 47-32 of this address.
This access is used only for symbols within text segment.

For symbols in data segment, any access must be made with an offset on
R2 register. However, it isn't simple to create this access during the
compilation (the symbol's section isn't set yet). Therefore, the
modification occurs during the link tool. It will modify the previous
4 instructions to use R2 and its symbol offset to create the correct
access.
During this process, we loose two instructions. This will be fixed
later.

LBRAPIC value was missing in anames9.go.

Update: #25893

Change-Id: Iab35f4b01afaec5bee5daaa32f0b2273f9a00fdf