cmd: fix symbols addressing for aix/ppc64
This commit changes the code generated for addressing symbols on AIX
operating system.

On linux/ppc64, symbols addresses are 32 bits but on aix/ppc64, they
are up to 64 bits. Therefore, this requires two more instructions and a
new relocation type to generate the larger address.
This addressing is used only for symbols within text segment.

For symbols in data segment, any access must be made with an offset on
R2 register. However, for some symbols, it is not known until link time
if they are in data segment or text segment. Therefore, some addressing
instructions must be changed at link time.

During this process, as the offset for data symbols are only up to
32 bits, the two newly added instructions, for 64 bits, are useless.
This will be fixed later.

LBRAPIC value was missing in anames9.go as an operand type.

Update: #25893

Change-Id: Iab35f4b01afaec5bee5daaa32f0b2273f9a00fdf