runtime: port assembly files for AIX
This commit adds the change on asm_ppc64.s and tls_ppc64.s files for AIX
operating system.

On AIX, the TOC (ie R2) is the same during the whole Golang execution.
It only changes during the call of a C function (cf syscalls).
Moreover, g must always be saved on the tls as syscalls are made with
C functions.

Some modifications on asm_ppc64.s are linked with the new stack layout.

It also removes a useless part in asmcgocall which was done twice.

Change-Id: Ie037ab73da00562bb978f2d0f17fcdabd4a40aa2