runtime: handle 64bits addresses for AIX
This commit allows the runtime to handle 64bits addresses returned by
mmap syscall on AIX.

Mmap syscall returns addresses on 59bits on AIX. But the Arena
implementation only allows addresses with less than 48 bits.
Therefore, AIX uses the value arenaBaseOffsetAIX to shift a pointer
address back on 48 bits with the functions arenaIndex and arenaBase.

The arenaBaseOffset value must be split in two differents values, one
for AMD and one for AIX. This is mandatory because a uintptr can't be
negative.

Update: #25893

Change-Id: Iea72e8a944d10d4f00be915785e33ae82dd6329e