cmd/compile: add option for user-supplied list of file/line/column specials
DO NOT SUBMIT

This is partly for estimating escape analysis upside from tracking
interface methods called on parameters, and partly to play with
a mechanism we might use for feedback-directed optimization.

This is not in a terribly clean state, but submitting it puts
it someplace safer than my local sandbox.

The code to capture this information for benchmarks is not
correct; it reports wrong (too small) information.

Change-Id: I0d264aa1b74b40d1bd7d6337ce9a4340bc0747ba