Update patch set 1
Patch Set 1:

The trybot error seems to be real, and it needs to be fixed so that we don't break all the trybots when this change is submitted.  You should be able to recreate the problem by going to the cmd/vet/all directory and running `go run main.go -p aix/ppc64`.  That command needs to succeed.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>