Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> I'm afraid I'm really confused by this change. Partly, I don't understand at all how the AIX address space is laid out. I looked at some IBM documents and didn't find anything indicating that the address space starts at 0xa0 << 52. And if mmap returns 59 bit addresses, then it seems like the arena index needs to cover 59 bits or maybe 64 bits (which it's completely capable of doing). Is there a reason not to just use a full 64 bit arena index?
> 
> Also, I don't understand why this is introducing a second offset. Why not define arenaBaseOffset on AIX to be ~uintptr(0xa0 << 52) + 1? (Though no offset is necessary if you use the full 64 bit arena index.)

Mmap on AIX always returns above 0xa0<<52 on AIX 7.2 even if the documentation isn't explicit about it.  

This code was made 4/5 months ago, and I wasn't able to get the arena working on 60 bits. The compiler was blocking me on some constants because of overflows or because objects were 'larger than address space'. At that time, I had found this workaround. But it seems that I can make it works if I adapt some constants in this file and inside cmd/compile.  

However, I've also chosen to keep the arena on 48 bits with an offset to don't have a too large arena which can affect performances. Because if L1 and L2 are too wide, it will reduce performances a lot isn't it ? How do I choose values like arenaSize, L1 and L2 if I set up a 60 bits arena ? Somethink like the following seems ok ?
  	//       Platform  Addr bits  Arena size  L1 entries  L2 size
	// --------------  ---------  ----------  ----------  -------
	//     aix/64-bit         60        64MB         4KB     4MB

I have still some experimentations to do but yes it might be better with a arena on 60 bits. 

Just for the offset, it's just because amd needs it as a positive value and aix as a negative one. But uintptr can't be negative so I can't have it in the same constant without having an if inside arenaIndex/arenaBase functions. (At least, I didn't found how).

Patch-set: 4
Reviewer: Gerrit User 27994 <27994@62eb7196-b449-3ce5-99f1-c037f21e1705>