runtime: scavenge large spans before heap growth
This change scavenges the largest spans before growing the heap for
physical pages to "make up" for the newly-mapped space which,
presumably, will be touched.

This approach to scavenging helps reduce the RSS of an application by
returning some of the memory left over as a result of fragmentation
back to the OS more eagerly.

Fixes #14045.

Change-Id: Iab60790be05935865fc71f793cb9323ab00a18bd