runtime: mark a freshly sysAlloc'd span scavenged
For new pages that are created via sysAlloc and then placed into a
span, we can consider those scavenged. On all platforms, sysAlloc only
maps in the pages and relies on OS demand paging to fill them in.
This way, we can also mark space we actually want to use with sysUsed,
improving the granularity of the bookkeeping a little bit.

Change-Id: I0718fca7f253d3f31b964202cd85079a20d50c41