Update patch set 5
Patch Set 5:

How close do we want to get to parameter expansion like POSIX shells do it? If we want to be really close, the code and API would be way more complex, and there are already third-party packages outside of the standard library that handle all that complexity.

Not saying that we should reject this, but we likely want to draw a line somewhere. It's not clear to me that we want this API to be much more advanced than what it already is.

Patch-set: 5
CC: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>