Update patch set 1
Patch Set 1:

It would be nice if when presented with FMOV[DS] $0, Fx the assembler just did 'the right thing'. That way we wouldn't need to introduce new rules. (s390x does this, calling the LOAD ZERO instruction if the FMOVD immediate is zero.)

Is there any reason not to defer this lowering to the assembler?

Patch-set: 1
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>