Update patch set 1
Patch Set 1: Code-Review-1

“I am working on a TLS server program, which issues new TLS certificates
on demand. The new certificates will be added into tls.Config.Certificates.
BuildNameToCertificate will be called to refresh the name table afterwards.”

The actual motivation for this change isn’t safe. You can’t modify the Certificates field (or call BuildNameToCertificate) while the server is running, it’s a race condition.

The only way this would be safe is to use GetConfigForClient and return a new tls.Config everytime it’s modified, in which case you’d be better to use GetCertificate. But because you can’t know when a connection has finished using the Config, there is actually no safe time to ever manipulate a tls.Config after it’s been used.

The for loop in BuildNameToCertificate only iterates once per-Certificate in the slice, meaning this change only improves the performance in a racey and unsafe use case. This shouldn’t be merged.

Patch-set: 1
Reviewer: Gerrit User 20190 <20190@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1