Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > I could also add a test in test/codegen, but not sure where it should go, any suggestions?
> 
> Yeah, that would be good. Maybe test/codegen/floats.go? Either that or we could add a new consts.go.

Oh, a codegen test is not going to work for this since the change to generate xxlxor was done in the assembler, and these tests look at the Go asm to decide if the correct code is being generated. The Go asm before and after my change is FMOVD (0.0), F1.

Patch-set: 5
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>