Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > The actual motivation for this change isn’t safe. You can’t modify the Certificates field (or call BuildNameToCertificate) while the server is running, it’s a race condition.
> > 
> > The only way this would be safe is to use GetConfigForClient and return a new tls.Config everytime it’s modified, in which case you’d be better to use GetCertificate. But because you can’t know when a connection has finished using the Config, there is actually no safe time to ever manipulate a tls.Config after it’s been used.
> > 
> > The for loop in BuildNameToCertificate only iterates once per-Certificate in the slice, meaning this change only improves the performance in a racey and unsafe use case. This shouldn’t be merged.

I can see a safe way to do this by combining a master (unused) Config on which you call BuildNameToCertificate with a lock, and GetConfigForClient in which you call Config.Clone.

The other reason this is useful is that if for any other reason you have parsed the Leaf before we should use it.

> I am using GetCertificate() with a lock, similar to what you described. However I still need BuildNameToCertificate() to build newly generated certificates. The existing certificates are alway re-built because the leaf certs are not cached. It would be great that BuildNameToCertificate() becomes independent from tls.Config, and let users to maintain an out-of-box name map.

I don't see how this would be useful if you are using GetCertificate, which ignores Certificates, did you mean GetConfigForClient? Note though that GetCertificate does let you maintain an out-of-box name map, so you might want to switch to that callback.

Patch-set: 1
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>